virt: add kernel memory area
This patch is the first in series of patches that split OP-TEE RW memory
into two regions: kernel memory and TEE memory. Kernel memory will
be always mapped and it will be used for to store all data that is
vital for OP-TEE core and is not bound to virtual guests.

TEE memory is a memory that holds data specific for certain guest.
There will be TEE memory bank for every guest and it will be mapped
into OP-TEE address space only during call from that guest.

This patch adds kernel memory and moves stacks into it. Also
it provides __kbss and __kdata macros, so one can easily set right
section for a variable.

Signed-off-by: Volodymyr Babchuk <vlad.babchuk@gmail.com>