plat: rcar: add ability to determine SoC type in runtime
This patch allows to determine SoC type and revision in runtime. This
is the first step in unifying all RCAR Gen3 flavors into one.

This code is loosely based on Renesas OP-TEE fork and on Renesas
ARM TF fork.

Depending on compile options (if ASLR enabled or not), this code can be
executed very early, when MMU is still disabled, so it checks MMU state
before accessing registers that hold SoC revision.

Signed-off-by: Volodymyr Babchuk <volodymyr_babchuk@epam.com>